refactor: rename aboutus folder
