+++ src/App.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import Home from './pages/HomePage/HomePage';
 import NotFound from './pages/404Page/404Page';
-import AboutUs from './pages/AboutUs/AboutUs';
+import AboutUs from './pages/AboutUsPage/AboutUs';
 class App extends React.Component<object, object> {
   render() {
     return (
